Leet Code Problem 290 Solved
Given a pattern and a string, our task is to check whether the pattern
and the string follow the same order. If they do, return true;
otherwise, return false. For example, with the pattern 'abba' and the
string 'dog cat cat dog,' it returns true. However, for the pattern
'abba' and the string 'dog cat cat fish,' it returns false.

In this solution, I utilized the JavaScript built-in Map object. First,
I converted the string into an array. Then, I iterated through the
pattern and the array simultaneously, checking if the current pattern
character already had a corresponding value in the Map. If it did, I
compared it with the current string word. If there was a mismatch or the
pattern character didn't exist in the Map but the string word did, I
returned false. Otherwise, I stored the pattern character and the string
word in the Map for future reference.

This approach allowed me to efficiently track the relationship between
pattern characters and string words, ensuring that the order was
maintained throughout the comparison process